feat(typechecker): initial work on type checking
The type checker is integrated with the CLI.

- Fix tokenizer bug: binary ops were not correctly parsed if operands were
of different types (i.e. INT and FLOAT). The tokenizer recognized a
float as being a token that starts with a number and contains a decimal
before a semicolon. This alogrithm is naive as multiple numeric tokens can
be on a single line before a semicolon (e.g. binary op). Furthermore, a
polynomial with mixed types, which is not permitted by the type checker,
but can still be written by the programmer, results in numeric tokens that
should be INTEGER being incorrectly classified as FLOAT. If the original
algorithm detects a decimal point *in another operand* and before a semi-
colon, it is possible that the digit that was encountered is considered a
FLOAT.
- Fix parser bug: polynomial expressions were not correctly parsed. The
parser did not continue pass the first operator. Therefore subsequent
operators were not processed resulting in false parser errors.